requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
gradio==4.5.0
python-dotenv==1.0.0
pydantic==2.4.2

# LLM and NLP
openai==1.3.5
langchain==0.0.335
sentence-transformers==2.2.2

# Document processing
pypdf==3.15.1
python-docx==0.8.11
faiss-cpu
#faiss-cpu==1.7.4

# Image processing
pillow==10.0.0

# Utilities
pytest==7.4.3
black==23.10.1
isort==5.12.0

PyPDF2